+++ src/emulator/databaseEmulator.ts
@@ -4,11 +4,11 @@ import * as fs from "fs";
 import * as path from "path";
 
 import * as api from "../api";
-import * as utils from "../utils";
 import * as downloadableEmulators from "./downloadableEmulators";
 import { EmulatorInfo, EmulatorInstance, Emulators } from "../emulator/types";
 import { Constants } from "./constants";
 import { EmulatorRegistry } from "./registry";
+import { EmulatorLogger } from "./emulatorLogger";
 
 export interface DatabaseEmulatorArgs {
   port?: number;
@@ -21,7 +21,8 @@ export interface DatabaseEmulatorArgs {
 }
 
 export class DatabaseEmulator implements EmulatorInstance {
-  rulesWatcher?: chokidar.FSWatcher;
+  private rulesWatcher?: chokidar.FSWatcher;
+  private logger = EmulatorLogger.forEmulator(Emulators.DATABASE);
 
   constructor(private args: DatabaseEmulatorArgs) {}
 
@@ -33,6 +34,16 @@ export class DatabaseEmulator implements EmulatorInstance {
     }
     if (this.args.rules) {
       for (const c of this.args.rules) {
+        if (!c.instance) {
+          this.logger.log("DEBUG", `args.rules=${JSON.stringify(this.args.rules)}`);
+          this.logger.logLabeled(
+            "WARN_ONCE",
+            "database",
+            "Could not determine your Realtime Database instance name, so rules hot reloading is disabled."
+          );
+          continue;
+        }
+
         const rulesPath = c.rules;
         this.rulesWatcher = chokidar.watch(rulesPath, { persistent: true, ignoreInitial: true });
         this.rulesWatcher.on("change", async (event, stats) => {
@@ -41,17 +52,18 @@ export class DatabaseEmulator implements EmulatorInstance {
           // Adding a small delay prevents that at very little cost.
           await new Promise((res) => setTimeout(res, 5));
 
-          utils.logLabeledBullet(
+          this.logger.logLabeled(
+            "BULLET",
             "database",
             `Change detected, updating rules for ${c.instance}...`
           );
           const newContent = fs.readFileSync(rulesPath, "utf8").toString();
           try {
             await this.updateRules(c.instance, newContent);
-            utils.logLabeledSuccess("database", "Rules updated.");
+            this.logger.logLabeled("SUCCESS", "database", "Rules updated.");
           } catch (e) {
-            utils.logWarning(this.prettyPrintRulesError(rulesPath, e));
-            utils.logWarning("Failed to update rules");
+            this.logger.logLabeled("WARN", "database", this.prettyPrintRulesError(rulesPath, e));
+            this.logger.logLabeled("WARN", "database", "Failed to update rules");
           }
         });
       }
@@ -60,8 +72,18 @@ export class DatabaseEmulator implements EmulatorInstance {
     return downloadableEmulators.start(Emulators.DATABASE, this.args);
   }
 
-  connect(): Promise<void> {
-    return Promise.resolve();
+  async connect(): Promise<void> {
+    // The chokidar watcher will handle updating rules but we need to set the initial ruleset for
+    // each namespace here.
+    if (this.args.rules) {
+      for (const c of this.args.rules) {
+        if (!c.instance) {
+          continue;
+        }
+
+        await this.updateRules(c.instance, fs.readFileSync(c.rules, "utf8").toString());
+      }
+    }
   }
 
   stop(): Promise<void> {
@@ -73,8 +95,10 @@ export class DatabaseEmulator implements EmulatorInstance {
     const port = this.args.port || Constants.getDefaultPort(Emulators.DATABASE);
 
     return {
+      name: this.getName(),
       host,
       port,
+      pid: downloadableEmulators.getPID(Emulators.DATABASE),
     };
   }
 
@@ -86,7 +110,7 @@ export class DatabaseEmulator implements EmulatorInstance {
     const { host, port } = this.getInfo();
     try {
       await api.request("PUT", `/.settings/rules.json?ns=${instance}`, {
-        origin: `http://${host}:${[port]}`,
+        origin: `http://${host}:${port}`,
         headers: { Authorization: "Bearer owner" },
         data: content,
         json: false,